Create EDA
In this unit the focus is on exploratory data analysis. EDA is performed in order to define and refine the selection of feature variables that will be used for machine learning. A data scientist has to become familiar with the data set they are working on, and understand the features of their data. This stage is important in getting the firm feature set needed for supervised and unsupervised machine learning.